+++ packages/temen/src/index.ts
@@ -1,16 +1,5 @@
 /** @module Temen */
 export { default as delay } from './delay';
-export {
-  createLocalStorageChangeEvent,
-  LocalStorageEventPayload,
-} from './localStorage/localStorageEvent';
-export {
-  setLocalStorageItem,
-  getLocalStorageItem,
-  removeLocalStorageItem,
-  popLocalStorageItem,
-  clearLocalStorage,
-} from './localStorage';
 export { default as commaizeNumber } from './commaizeNumber';
 export { default as getObjectKeys } from './getObjectKeys';
 export * from './is';
@@ -19,11 +8,8 @@ export * from './promise';
 export * from './date';
 export { default as getArrayFromCount } from './getArrayFromCount';
 export * from './createFromArray';
-export * from './copyToClipboard';
 export * from './svg';
-export * from './fetch';
 export * from './uuid';
-export * from './cookies';
 export * from './functions';
 export * from './models/utils';
 export { default as chunk } from './chunk';
@@ -45,4 +31,5 @@ export * from './sum';
 export * from './max';
 export * from './min';
 export * from './average';
+export * from './drop';
 export * from './throttle';