+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.21](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/icons@1.2.20...@lubycon/icons@1.2.21) (2021-12-19)
+
+**Note:** Version bump only for package @lubycon/icons
+
+
+
+
+
+## [1.2.20](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/icons@1.2.19...@lubycon/icons@1.2.20) (2021-12-19)
+
+**Note:** Version bump only for package @lubycon/icons
+
+
+
+
+
 ## [1.2.19](https://github.com/Lubycon/lubycon-frontend-libraries/compare/@lubycon/icons@1.2.18...@lubycon/icons@1.2.19) (2021-12-16)
 
 **Note:** Version bump only for package @lubycon/icons